Step 2 move register code to library
