+++ platform/kube/BUILD
@@ -8,13 +8,14 @@ go_library(
         "controller.go",
         "conversion.go",
         "queue.go",
+        "register.go",
     ],
     visibility = ["//visibility:public"],
     deps = [
         "//model:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
-        "@io_istio_api//:go_default_library",
+        "@io_istio_api//proxy/v1/config:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
