+++ platform/kube/inject/BUILD
@@ -14,7 +14,7 @@ go_library(
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
-        "@io_istio_api//:go_default_library",
+        "@io_istio_api//proxy/v1/config:go_default_library",
         "@io_k8s_api//apps/v1beta1:go_default_library",
         "@io_k8s_api//batch/v1:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
@@ -46,7 +46,7 @@ go_test(
         "//proxy:go_default_library",
         "//test/util:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
-        "@io_istio_api//:go_default_library",
+        "@io_istio_api//proxy/v1/config:go_default_library",
         "@io_k8s_api//apps/v1beta1:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
